requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google==3.0.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.0
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9
soupsieve==2.6
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
uWSGI==2.0.28
